:truck: rename: 회원가입 완료 페이지 컴포넌트 분리
